requirements.txt
aiogram-3.1.1
openaiAPI
i18nice-0.12.0
